pyproject.toml
[tool.poetry]
name = "aarons-toolkit"
version = "0.1.0"
description = "a decentralised scraper designed to help researchers and students liberate academic research locked behind paywalls."
authors = ["Aarons KIT"]
readme = "README.md"
packages = [{include = "aarons_toolkit"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
selenium = "^4.8.3"
webdriver-manager = "^3.8.5"
SpeechRecognition = "^3.10.0"
speedtest-cli = "^2.1.3"
emoji = "^2.2.0"
boto3 = "^1.26.107"
termcolor = "^2.2.0"
py-algorand-sdk = "^2.1.2"
pyFLAC = "^2.1.0"
auto-py-to-exe = "^2.35.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
